requirements.txt
absl-py==1.4.0
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
cachetools==4.2.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clldutils==3.19.0
cmake==3.26.1
colorama==0.4.6
colorlog==6.7.0
cPython==0.0.6
cryptography==40.0.1
csvw==3.1.3
cycler==0.11.0
Cython==0.29.21
decorator==5.1.1
Deprecated==1.2.13
dnspython==2.3.0
filelock==3.10.7
fugashi==1.2.1
future==0.18.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
grpcio==1.53.0
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipadic==1.0.0
isodate==0.6.1
jaconv==0.3.4
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
keyring==23.13.1
kiwisolver==1.4.4
language-tags==1.2.0
librosa==0.8.0
lit==16.0.0
llvmlite==0.39.1
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.3.1
mecab-python3==1.0.6
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.18.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.0
pexpect==4.8.0
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
pooch==1.7.0
protobuf==3.20.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pykakasi==2.2.1
pylatexenc==2.10
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
rdflib==6.3.2
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.2.2
scipy==1.5.2
SecretStorage==3.3.3
segments==2.2.1
six==1.16.0
soundfile==0.12.1
ssh-run==3.1.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
threadpoolctl==3.1.0
torch==2.0.0
torchaudio==2.0.1
torchvision==0.15.1
triton==2.0.0
typing-extensions==4.5.0
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
